Relocate the spool-side inventory machinery into one core module
Everything a spool asks of an attached inventory -- row-epoch
resolution, channel placement, selection and splitting along the fiber,
the lazy InventoryRef holder, and the shared projection primitives --
now lives in dascore/core/_spool_inventory.py. dascore/proc/inventory.py
keeps only Patch.enrich and its patch-facing helpers, importing the
shared primitives, so the proc namespace no longer exports six spool
planning functions and no proc module reaches into dascore.io.

spool.py imports the machinery at module top, replacing eleven deferred
proc imports, and its selection paths flatten: one _classify_query feeds
both select and unselect, so unselect no longer re-implements the
classification select uses. The selector-spec helpers join
resolve_selector_namespaces in utils/pd.py, where the spec's forms are
already parsed.
